Extract PatternContentSanitizer and restore private visibility on SavePatternsToTheme
Move pure text-processing functions (sanitizePatternContent, generateLoremIpsum,
countWords) into a dedicated PatternContentSanitizer class. Revert generateSlug,
formatPatternFile, and findExistingFileForPost to private. Remove resetUsedSlugs.
Rewrite tests to exercise the public API (savePatternToTheme, handleRestRequest).
